Split endpoints into corresponding routers
